Added uriNavigateRequest to editor and host opens folder, didn't test w/ files. Added link click handlers and hacky ways of using insert link from toolbar element, so many bugs with quill half of what I do is work around its bugs. Figured out a workaround for xml whitespace issue while sacrificing good html formatting which was the whole point of using the stupid thing. I should have made my own editor its becoming silly. Lots of weird things going on and just adding hacks to fix which is not good. But I'm fixing them but fixing one thing breaks another, progress is slow.

+++ Common/MonkeyPaste.Common.Wpf/Utils/Converters/MpWpfRtfToHtmlConverter.cs
@@ -329,7 +329,7 @@ namespace MonkeyPaste.Common.Wpf {
         }
 
         private static string GetFontSize(Span s) {
-            double fs = (double)s.FontSize;//new FontSizeConverter().ConvertFrom(s.FontSize+"pt");
+            double fs = (double)((int)s.FontSize);//new FontSizeConverter().ConvertFrom(s.FontSize+"pt");
             //MpWpfRtfDefaultProperties.Instance.AddFontSize(fs);
             return string.Format(@" style='font-size: {0}px;", fs);
         }
